# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ProductSubcategory {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  value: string;
}

export interface Product {
  id: number;
  name: string;
  sku: string;
  category: string;
  /** Optional product attributes such as size or color. */
  subcategories: ProductSubcategory[];
  costPrice: number;
  salePrice: number;
  stock: number;
  /** @nullable */
  imageUrl: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sku: string;
  /** @minLength 1 */
  category: string;
  subcategories: ProductSubcategory[];
  /** @minimum 0 */
  costPrice: number;
  /** @minimum 0 */
  salePrice: number;
  /** @minimum 0 */
  stock: number;
  /** @nullable */
  imageUrl: string | null;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  sku?: string;
  /** @minLength 1 */
  category?: string;
  subcategories?: ProductSubcategory[];
  /** @minimum 0 */
  costPrice?: number;
  /** @minimum 0 */
  salePrice?: number;
  /** @nullable */
  imageUrl?: string | null;
}

export type InventoryMovementType = typeof InventoryMovementType[keyof typeof InventoryMovementType];


export const InventoryMovementType = {
  purchase: 'purchase',
  sale: 'sale',
} as const;

export interface InventoryMovement {
  id: number;
  productId: number;
  productName: string;
  productSku: string;
  type: InventoryMovementType;
  quantity: number;
  unitPrice: number;
  total: number;
  stockAfter: number;
  /** @nullable */
  counterparty: string | null;
  /** @nullable */
  note: string | null;
  createdAt: string;
}

export type InventoryMovementInputType = typeof InventoryMovementInputType[keyof typeof InventoryMovementInputType];


export const InventoryMovementInputType = {
  purchase: 'purchase',
  sale: 'sale',
} as const;

export interface InventoryMovementInput {
  productId: number;
  type: InventoryMovementInputType;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  unitPrice: number;
  /** @maxLength 120 */
  counterparty?: string;
  /** @maxLength 300 */
  note?: string;
}

export interface ProductSummary {
  totalProducts: number;
  totalUnits: number;
  inventoryValue: number;
  projectedProfit: number;
  lowStockCount: number;
  categories: number;
}

export type AssistantChatMessageRole = typeof AssistantChatMessageRole[keyof typeof AssistantChatMessageRole];


export const AssistantChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface AssistantChatMessage {
  role: AssistantChatMessageRole;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  content: string;
}

export interface AssistantChatInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
  /** @maxItems 12 */
  history?: AssistantChatMessage[];
}

export interface AssistantChatResponse {
  reply: string;
}

export type ListProductsParams = {
search?: string;
category?: string;
};

export type ListMovementsParams = {
productId?: number;
type?: ListMovementsType;
};

export type ListMovementsType = typeof ListMovementsType[keyof typeof ListMovementsType];


export const ListMovementsType = {
  purchase: 'purchase',
  sale: 'sale',
} as const;