# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog operations
  - name: movements
    description: Inventory Kardex operations
  - name: assistant
    description: Inventory assistant operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Product deleted
        "404":
          description: Product not found
  /products/summary:
    get:
      operationId: getProductSummary
      tags: [products]
      summary: Get catalog summary
      responses:
        "200":
          description: Catalog summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSummary"
  /movements:
    get:
      operationId: listMovements
      tags: [movements]
      summary: List inventory movements
      parameters:
        - name: productId
          in: query
          required: false
          schema:
            type: number
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [purchase, sale]
      responses:
        "200":
          description: Inventory movement list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryMovement"
    post:
      operationId: createMovement
      tags: [movements]
      summary: Register an inventory movement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryMovementInput"
      responses:
        "201":
          description: Created inventory movement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryMovement"
        "400":
          description: Invalid movement or insufficient stock
        "404":
          description: Product not found
  /assistant/chat:
    post:
      operationId: chatWithInventoryAssistant
      tags: [assistant]
      summary: Ask the inventory assistant a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantChatInput"
      responses:
        "200":
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantChatResponse"
        "400":
          description: Invalid message
        "502":
          description: AI provider error
        "503":
          description: AI provider is not configured
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        sku:
          type: string
        category:
          type: string
        subcategories:
          type: array
          description: Optional product attributes such as size or color.
          items:
            $ref: "#/components/schemas/ProductSubcategory"
        costPrice:
          type: number
        salePrice:
          type: number
        stock:
          type: number
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, sku, category, subcategories, costPrice, salePrice, stock, imageUrl, createdAt, updatedAt]
    ProductSubcategory:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        value:
          type: string
          minLength: 1
      required: [name, value]
    ProductInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        sku:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        subcategories:
          type: array
          items:
            $ref: "#/components/schemas/ProductSubcategory"
        costPrice:
          type: number
          minimum: 0
        salePrice:
          type: number
          minimum: 0
        stock:
          type: number
          minimum: 0
        imageUrl:
          type: ["string", "null"]
      required: [name, sku, category, subcategories, costPrice, salePrice, stock, imageUrl]
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        sku:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        subcategories:
          type: array
          items:
            $ref: "#/components/schemas/ProductSubcategory"
        costPrice:
          type: number
          minimum: 0
        salePrice:
          type: number
          minimum: 0
        imageUrl:
          type: ["string", "null"]
    InventoryMovement:
      type: object
      properties:
        id:
          type: number
        productId:
          type: number
        productName:
          type: string
        productSku:
          type: string
        type:
          type: string
          enum: [purchase, sale]
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
        stockAfter:
          type: number
        counterparty:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, productId, productName, productSku, type, quantity, unitPrice, total, stockAfter, counterparty, note, createdAt]
    InventoryMovementInput:
      type: object
      properties:
        productId:
          type: number
        type:
          type: string
          enum: [purchase, sale]
        quantity:
          type: number
          minimum: 1
        unitPrice:
          type: number
          minimum: 0
        counterparty:
          type: string
          maxLength: 120
        note:
          type: string
          maxLength: 300
      required: [productId, type, quantity, unitPrice]
    ProductSummary:
      type: object
      properties:
        totalProducts:
          type: number
        totalUnits:
          type: number
        inventoryValue:
          type: number
        projectedProfit:
          type: number
        lowStockCount:
          type: number
        categories:
          type: number
      required: [totalProducts, totalUnits, inventoryValue, projectedProfit, lowStockCount, categories]
    AssistantChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
          maxLength: 4000
      required: [role, content]
    AssistantChatInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 2000
        history:
          type: array
          maxItems: 12
          items:
            $ref: "#/components/schemas/AssistantChatMessage"
      required: [message]
    AssistantChatResponse:
      type: object
      properties:
        reply:
          type: string
      required: [reply]