# lib/api-zod/src/generated/types/inventoryMovement.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryMovementType } from './inventoryMovementType';

export interface InventoryMovement {
  id: number;
  productId: number;
  productName: string;
  productSku: string;
  type: InventoryMovementType;
  quantity: number;
  unitPrice: number;
  total: number;
  stockAfter: number;
  /** @nullable */
  counterparty: string | null;
  /** @nullable */
  note: string | null;
  createdAt: Date;
}