# lib/api-zod/src/generated/types/inventoryMovementInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryMovementInputType } from './inventoryMovementInputType';

export interface InventoryMovementInput {
  productId: number;
  type: InventoryMovementInputType;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  unitPrice: number;
  /** @maxLength 120 */
  counterparty?: string;
  /** @maxLength 300 */
  note?: string;
}