# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductSubcategory } from './productSubcategory';

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  sku?: string;
  /** @minLength 1 */
  category?: string;
  subcategories?: ProductSubcategory[];
  /** @minimum 0 */
  costPrice?: number;
  /** @minimum 0 */
  salePrice?: number;
  /** @nullable */
  imageUrl?: string | null;
}